% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{family.ivm}
\alias{family.ivm}
\title{Model family for ivm}
\usage{
\method{family}{ivm}(object, ...)
}
\arguments{
\item{object}{ivm model object.}

\item{...}{other arguments}
}
\description{
Model family for ivm
}
