% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{find_instruments}
\alias{find_instruments}
\title{Internal function for identifying instruments}
\usage{
find_instruments(formula1, formula2)
}
\arguments{
\item{formula1}{model formula}

\item{formula2}{instrument formula}
}
\description{
Internal function for identifying instruments
}
