% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivglm.R
\name{iv.glm}
\alias{iv.glm}
\title{Estimate linear regression models with an instrumental variable.}
\usage{
iv.glm(formula, instrument_formula, data = NULL, family = binomial,
  link = "logit", ...)
}
\arguments{
\item{formula}{the formula of the linear regression, including variable}
}
\description{
This function allows you to estimate a two-stage least squares linear regression in one step.
}
\examples{
# Fake data
N <- 1000
z <- rnorm(N, 1, 1)
error <- rnorm(N, 1, 1)
x <- z + error + rnorm(N, 1, 1)
y <- x + error

Fit glm
fit_glm <- glm(formula = y ~ x, family = binomial(link = 'logit'))

Fit with iv
fit_2sls <- iv.glm(formula = y ~ instrument, instrument_formula = x ~ z, family = binomial, link = 'logit')

summary(fit_glm)
summary(fit_2sls)
}
\keyword{instrument}
