% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{model.frame.ivm}
\alias{model.frame.ivm}
\title{Model frame for ivm}
\usage{
\method{model.frame}{ivm}(formula, ...)
}
\arguments{
\item{formula}{ivm model object.}

\item{...}{other arguments}
}
\description{
Model frame for ivm
}
