% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{model.matrix.ivm}
\alias{model.matrix.ivm}
\title{Model matrix for ivm}
\usage{
\method{model.matrix}{ivm}(object, ...)
}
\arguments{
\item{object}{ivm model object.}

\item{...}{other arguments}
}
\description{
Model matrix for ivm
}
