% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{nobs.ivm}
\alias{nobs.ivm}
\title{Number of observations in ivm model}
\usage{
\method{nobs}{ivm}(object, ...)
}
\arguments{
\item{object}{ivm model object.}

\item{...}{other arguments}
}
\description{
Number of observations in ivm model
}
