% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{predict.ivm}
\alias{predict.ivm}
\title{Make predictions from ivm model}
\usage{
\method{predict}{ivm}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{ivm model object.}

\item{newdata}{(optional) new data for predictions.}

\item{...}{other arguments.}
}
\description{
Make predictions from ivm model
}
