% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{residuals.ivm}
\alias{residuals.ivm}
\title{Get residuals from ivm model}
\usage{
\method{residuals}{ivm}(object, ...)
}
\arguments{
\item{object}{ivm model object.}

\item{...}{other arguments}
}
\description{
Get residuals from ivm model
}
