% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.ivm}
\alias{summary.ivm}
\title{Model summary for ivm}
\usage{
\method{summary}{ivm}(object, ...)
}
\arguments{
\item{object}{ivm model object.}

\item{...}{other arguments}
}
\description{
Model summary for ivm
}
