% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{variable.names.ivm}
\alias{variable.names.ivm}
\title{Get variables in  ivm model}
\usage{
\method{variable.names}{ivm}(object, ...)
}
\arguments{
\item{object}{ivm model object.}

\item{...}{other arguments}
}
\description{
Get variables in  ivm model
}
