% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivglm.R
\name{iv.glm}
\alias{iv.glm}
\title{Estimate linear regression models with an instrumental variable.}
\usage{
iv.glm(model_formula, instrument_formula, data = NULL, family = binomial,
  link = "logit", ...)
}
\arguments{
\item{model_formula}{The formula of the main regression problem.  Use `instrument` for the variable that is the result of the `instrument_formula` function.}

\item{instrument_formula}{The formula for the first stage of the regression problem.  Estimate an endogenous variable from one (or more) instruments. If blank, defaults to regular glm.}
}
\description{
This function allows you to estimate a two-stage least squares linear regression in one step.
}
\examples{
# Fake data
N <- 1000
z <- rnorm(N, 1, 1)
error <- rnorm(N, 1, 1)
x <- z + error + rnorm(N, 1, 1)
y <- x + error

Fit glm
fit_glm <- glm( y ~ x, family = binomial(link = 'logit'))

Fit with iv
fit_iv <- iv.glm(y ~ x, x ~ z, family = binomial, link = 'logit')

summary(fit_glm)
summary(fit_iv)
}
\concept{The distribution family of the outcome variable.}
\keyword{instrument}
