% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ivlm.R
\name{iv.lm}
\alias{iv.lm}
\title{Estimate linear regression models with an instrumental variable.}
\usage{
iv.lm(formula, instrument_formula, data = NULL, ...)
}
\arguments{
\item{formula}{the formula of the linear regression}
}
\description{
This function allows you to estimate a two-stage least squares linear regression in one step.
}
\examples{
# Fake data
N <- 1000
z <- rnorm(N, 1, 1)
error <- rnorm(N, 1, 1)
x <- z + error + rnorm(N, 1, 1)
y <- x + error

# Fit OLS
fit_ols <- lm(formula = y ~ x)

# Fit 2SLS
fit_2sls <- iv.lm(formula = y ~ instrument, instrument_formula = x ~ z)

summary(fit_ols)
summary(fit_2sls)
}
\keyword{instrument}
