% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subscriptions.R
\docType{data}
\name{subscriptions}
\alias{subscriptions}
\title{Data on customer spending, subscriptions, and churn.}
\format{A data frame with 100000 rows and 5 variables}
\usage{
data(subscriptions)
}
\description{
A simulated dataset of customer spending used to demonstrate the functionality of the instruments package.
}
\details{
\itemize{
  \item baseline_spending. latent variable with customers' baseline spending rate.
  \item incentive.  binary variable indicating whether customer was offered an incentive.
  \item subscription. whether customer has a subscription.
  \item total_spending. total spending over a given period.
  \item churn. binary variable indicating whether customer has churned.
}
}
\keyword{datasets}
